[rename] : Provider 파일 경로 수정(auth.domain -> member.domain)
